requirements.txt
streamlit
tensorflow
scikit-learn
numpy
pillow
joblib